disable ODBC tests for unstable

+++ docker/integration-tests/run-integration-tests.sh
@@ -254,7 +254,10 @@ else  # full images (MSSQL only !)
 
         docker build $add_host_cmd --build-arg ORTHANC_TESTS_REVISION=$orthanc_tests_revision -f orthanc-tests/Dockerfile --target orthanc-tests -t orthanc-tests orthanc-tests
 
-        COMPOSE_FILE=docker-compose.odbc-sql-server.yml             docker compose down -v
-        COMPOSE_FILE=docker-compose.odbc-sql-server.yml             docker compose up --build --exit-code-from orthanc-tests --abort-on-container-exit
+        if [[ "$version" == "stable" ]]; then
+            # TODO: re-enable ODBC tests in unstable mode
+            COMPOSE_FILE=docker-compose.odbc-sql-server.yml             docker compose down -v
+            COMPOSE_FILE=docker-compose.odbc-sql-server.yml             docker compose up --build --exit-code-from orthanc-tests --abort-on-container-exit
+        fi
     fi
 fi
